requirements.txt
astropy>=4.3
matplotlib==3.0.1
numpy>=1.21
scipy>=1.1
tqdm>=4.62
scikit-image>=0.19
pandas>=1.3
PyYAML>=6.0

python-dateutil~=2.8.2